+++ bgee-dao-api/src/main/java/org/bgee/model/dao/api/expressiondata/rawdata/est/ESTDAO.java
@@ -69,8 +69,10 @@ public interface ESTDAO extends DAO<ESTDAO.Attribute> {
      *                          of a same filter and uses OR between filters.
      * @param offset            An {@code Integer} used to specify which row to start from retrieving data
      *                          in the result of a query. If null, retrieve data from the first row.
-     * @param limit             An {@code Integer} used to limit the number of rows returned in a query
+     * @param limit             A {@code Long} used to limit the number of rows returned in a query
      *                          result. If null, all results are returned.
+     *                          {@code Long} because sometimes the number of potential results
+     *                          can be very large.
      * @param attributes        A {@code Collection} of {@code Attribute}s to specify the information
      *                          to retrieve from the data source.
      * @return                  A {@code ESTTOResultSet} allowing to retrieve the
@@ -78,7 +80,7 @@ public interface ESTDAO extends DAO<ESTDAO.Attribute> {
      * @throws DAOException     If an error occurred while accessing the data source.
      */
     public ESTTOResultSet getESTs(Collection<DAORawDataFilter> rawDataFilters,
-            Integer offset, Integer limit, Collection<Attribute> attributes) throws DAOException;
+            Long offset, Integer limit, Collection<Attribute> attributes) throws DAOException;
 
     public interface ESTTOResultSet extends DAOResultSet<ESTTO> {}
