+++ bgee-dao-api/src/main/java/org/bgee/model/dao/api/expressiondata/rawdata/insitu/InSituEvidenceDAO.java
@@ -73,8 +73,9 @@ public interface InSituEvidenceDAO extends DAO<InSituEvidenceDAO.Attribute> {
      * @param rawDatafilters    A {@code Collection} of {@code DAORawDataFilter} allowing to filter which
      *                          evidence to retrieve. The query uses AND between elements of a same filter and
      *                          uses OR between filters.
-     * @param offset            An {@code Integer} used to specify which row to start from retrieving data
+     * @param offset            A {@code Long} used to specify which row to start from retrieving data
      *                          in the result of a query. If null, retrieve data from the first row.
+     *                          {@code Long} because sometimes the number of results can be very large.
      * @param limit             An {@code Integer} used to limit the number of rows returned in a query
      *                          result. If null, all results are returned.
      * @param attributes        A {@code Collection} of {@code Attribute}s to specify the information
@@ -84,7 +85,7 @@ public interface InSituEvidenceDAO extends DAO<InSituEvidenceDAO.Attribute> {
      * @throws DAOException     If an error occurred while accessing the data source.
      */
     public InSituEvidenceTOResultSet getInSituEvidences(Collection<DAORawDataFilter> rawDatafilters,
-            Integer offset, Integer limit, Collection<Attribute> attributes) throws DAOException;
+            Long offset, Integer limit, Collection<Attribute> attributes) throws DAOException;
 
     public interface InSituEvidenceTOResultSet extends DAOResultSet<InSituEvidenceTO> {}
