+++ bgee-dao-api/src/main/java/org/bgee/model/dao/api/expressiondata/rawdata/microarray/AffymetrixChipDAO.java
@@ -80,8 +80,10 @@ public interface AffymetrixChipDAO extends DAO<AffymetrixChipDAO.Attribute> {
      * @param rawDatafilters    A {@code Collection} of {@code DAORawDataFilter} allowing to filter which
      *                          chips to retrieve. The query uses AND between elements of a same filter and
      *                          uses OR between filters.
-     * @param offset            An {@code Integer} used to specify which row to start from retrieving data
+     * @param offset            A {@code Long} used to specify which row to start from retrieving data
      *                          in the result of a query. If null, retrieve data from the first row.
+     *                          {@code Long} because sometimes the number of potential results
+     *                          can be very large.
      * @param limit             An {@code Integer} used to limit the number of rows returned in a query
      *                          result. If null, all results are returned.
      * @param attributes        A {@code Collection} of {@code Attribute}s to specify the information
@@ -91,7 +93,7 @@ public interface AffymetrixChipDAO extends DAO<AffymetrixChipDAO.Attribute> {
      * @throws DAOException     If an error occurred while accessing the data source.
      */
     public AffymetrixChipTOResultSet getAffymetrixChips(Collection<DAORawDataFilter> rawDatafilters,
-            Integer offset, Integer limit, Collection<Attribute> attributes) throws DAOException;
+            Long offset, Integer limit, Collection<Attribute> attributes) throws DAOException;
 
     /**
      * {@code DAOResultSet} for {@code AffymetrixChipTO}s
