+++ bgee-dao-api/src/main/java/org/bgee/model/dao/api/expressiondata/rawdata/microarray/MicroarrayExperimentDAO.java
@@ -59,9 +59,11 @@ public interface MicroarrayExperimentDAO extends DAO<MicroarrayExperimentDAO.Att
      * @param rawDataFilters    A {@code Collection} of {@code DAORawDataFilter} allowing to specify
      *                          how to filter experiments to retrieve. The query uses AND between
      *                          elements of a same filter and uses OR between filters.
-     * @param offset            An {@code Integer} used to specify which row to start from retrieving data
+     * @param offset            A {@code Long} used to specify which row to start from retrieving data
      *                          in the result of a query. If null, retrieve data from the first row. If
-     *                          not null, a limit should be also provided
+     *                          not null, a limit should be also provided.
+     *                          {@code Long} because sometimes the number of potential results
+     *                          can be very large.
      * @param limit             An {@code Integer} used to limit the number of rows returned in a query
      *                          result. If null, all results are returned.
      * @param attributes        A {@code Collection} of {@code Attribute}s to specify the information
@@ -71,7 +73,7 @@ public interface MicroarrayExperimentDAO extends DAO<MicroarrayExperimentDAO.Att
      * @throws DAOException     If an error occurred while accessing the data source.
      */
     public MicroarrayExperimentTOResultSet getExperiments(Collection<DAORawDataFilter> rawDataFilters,
-            Integer offset, Integer limit, Collection<Attribute> attributes) throws DAOException;
+            Long offset, Integer limit, Collection<Attribute> attributes) throws DAOException;
 
 	/**
      * {@code DAOResultSet} for {@code MicroarrayExperimentTO}s
