+++ bgee-dao-api/src/main/java/org/bgee/model/dao/api/expressiondata/rawdata/rnaseq/RNASeqLibraryDAO.java
@@ -79,8 +79,10 @@ public interface RNASeqLibraryDAO extends DAO<RNASeqLibraryDAO.Attribute> {
      *                          If <strong>true</strong> only single-cell RNA-Seq are retrieved.
      *                          If <strong>false</strong> only bulk RNA-Seq are retrieved.
      *                          If <strong>null</strong> all RNA-Seq are retrieved.
-     * @param offset            An {@code Integer} used to specify which row to start from retrieving data
+     * @param offset            A {@code Long} used to specify which row to start from retrieving data
      *                          in the result of a query. If null, retrieve data from the first row.
+     *                          {@code Long} because sometimes the number of potential results
+     *                          can be very large.
      * @param limit             An {@code Integer} used to limit the number of rows returned in a query
      *                          result. If null, all results are returned.
      * @param attributes        A {@code Collection} of {@code Attribute}s to specify the information
@@ -91,7 +93,7 @@ public interface RNASeqLibraryDAO extends DAO<RNASeqLibraryDAO.Attribute> {
      * @throws DAOException     If an error occurred when accessing the data source.
      */
     public RNASeqLibraryTOResultSet getRnaSeqLibrary(Collection<DAORawDataFilter> rawDataFilters,
-            Boolean isSingleCell, Integer offset, Integer limit,
+            Boolean isSingleCell, Long offset, Integer limit,
             Collection<Attribute> attributes) throws DAOException;
 
     /**
