+++ bgee-dao-api/src/main/java/org/bgee/model/dao/api/expressiondata/rawdata/rnaseq/RNASeqResultAnnotatedSampleDAO.java
@@ -83,9 +83,11 @@ public interface RNASeqResultAnnotatedSampleDAO extends DAO<RNASeqResultAnnotate
      *                          If <strong>true</strong> only single-cell RNA-Seq are retrieved.
      *                          If <strong>false</strong> only bulk RNA-Seq are retrieved.
      *                          If <strong>null</strong> all RNA-Seq are retrieved.
-     * @param offset            An {@code Integer} used to specify which row to start from retrieving data
+     * @param offset            A {@code Long} used to specify which row to start from retrieving data
      *                          in the result of a query. If null, retrieve data from the first row. If
-     *                          not null, a limit should be also provided
+     *                          not null, a limit should be also provided.
+     *                          {@code Long} because sometimes the number of potential results
+     *                          can be very large.
      * @param limit             An {@code Integer} used to limit the number of rows returned in a query
      *                          result. If null, all results are returned.
      * @param attributes        A {@code Collection} of {@code Attribute}s to specify the information
@@ -96,7 +98,7 @@ public interface RNASeqResultAnnotatedSampleDAO extends DAO<RNASeqResultAnnotate
      */
     public RNASeqResultAnnotatedSampleTOResultSet getResultAnnotatedSamples(
             Collection<DAORawDataFilter> rawDataFilters, Boolean isSingleCell,
-            Integer offset, Integer limit, Collection<Attribute> attributes) throws DAOException;
+            Long offset, Integer limit, Collection<Attribute> attributes) throws DAOException;
 
     /**
      * {@code DAOResultSet} for {@code RNASeqExperimentTO}s
