+++ bgee-dao-sql/src/main/java/org/bgee/model/dao/mysql/expressiondata/rawdata/insitu/MySQLInSituEvidenceDAO.java
@@ -70,7 +70,7 @@ implements InSituEvidenceDAO{
 
     @Override
     public InSituEvidenceTOResultSet getInSituEvidences(Collection<DAORawDataFilter> rawDataFilters,
-            Integer offset, Integer limit, Collection<InSituEvidenceDAO.Attribute> attrs)
+            Long offset, Integer limit, Collection<InSituEvidenceDAO.Attribute> attrs)
             throws DAOException {
         log.traceEntry("{}, {}, {}, {}", rawDataFilters, offset, limit, attrs);
         checkOffsetAndLimit(offset, limit);
