+++ bgee-dao-sql/src/main/java/org/bgee/model/dao/mysql/expressiondata/rawdata/insitu/MySQLInSituExperimentDAO.java
@@ -34,7 +34,7 @@ implements InSituExperimentDAO{
 
     @Override
     public InSituExperimentTOResultSet getInSituExperiments(Collection<DAORawDataFilter> rawDataFilters,
-            Integer offset, Integer limit, Collection<InSituExperimentDAO.Attribute> attrs)
+            Long offset, Integer limit, Collection<InSituExperimentDAO.Attribute> attrs)
             throws DAOException {
         log.traceEntry("{}, {}, {}, {}", rawDataFilters, offset, limit, attrs);
         checkOffsetAndLimit(offset, limit);
