+++ bgee-dao-sql/src/main/java/org/bgee/model/dao/mysql/expressiondata/rawdata/rnaseq/MySQLRNASeqExperimentDAO.java
@@ -34,7 +34,7 @@ implements RNASeqExperimentDAO{
 
     @Override
     public RNASeqExperimentTOResultSet getExperiments(Collection<DAORawDataFilter> rawDataFilters,
-            Boolean isSingleCell, Integer offset, Integer limit,
+            Boolean isSingleCell, Long offset, Integer limit,
             Collection<RNASeqExperimentDAO.Attribute> attrs) throws DAOException {
         log.traceEntry("{}, {}, {}, {}, {}", rawDataFilters, isSingleCell, offset, limit, attrs);
